go/runtime/host: Move quotePolicyProvider to provisioner
Logic should not be nested in the common packages.
