+++ go/runtime/host/provisioner/provisioner.go
@@ -1,16 +1,22 @@
 package provisioner
 
 import (
+	"context"
 	"fmt"
 
+	"github.com/oasisprotocol/oasis-core/go/common"
+	"github.com/oasisprotocol/oasis-core/go/common/cbor"
 	"github.com/oasisprotocol/oasis-core/go/common/identity"
+	"github.com/oasisprotocol/oasis-core/go/common/node"
 	"github.com/oasisprotocol/oasis-core/go/common/persistent"
 	"github.com/oasisprotocol/oasis-core/go/common/sgx/pcs"
+	sgxQuote "github.com/oasisprotocol/oasis-core/go/common/sgx/quote"
 	"github.com/oasisprotocol/oasis-core/go/common/version"
 	"github.com/oasisprotocol/oasis-core/go/config"
 	consensus "github.com/oasisprotocol/oasis-core/go/consensus/api"
 	genesisAPI "github.com/oasisprotocol/oasis-core/go/genesis/api"
 	cmdFlags "github.com/oasisprotocol/oasis-core/go/oasis-node/cmd/common/flags"
+	registry "github.com/oasisprotocol/oasis-core/go/registry/api"
 	"github.com/oasisprotocol/oasis-core/go/runtime/bundle/component"
 	rtConfig "github.com/oasisprotocol/oasis-core/go/runtime/config"
 	runtimeHost "github.com/oasisprotocol/oasis-core/go/runtime/host"
@@ -48,7 +54,7 @@ func New(
 		return nil, err
 	}
 
-	policyProvider := sgxCommon.NewQuotePolicyProvider(consensus)
+	policyProvider := &quotePolicyProvider{consensus}
 
 	// Create runtime provisioner.
 	return createProvisioner(dataDir, commonStore, identity, hostInfo, qs, policyProvider)
@@ -195,4 +201,27 @@ func createProvisioner(
 	provisioner := hostComposite.NewProvisioner(provisioners)
 
 	return provisioner, nil
+}
+
+type quotePolicyProvider struct {
+	cs consensus.Service
+}
+
+func (p *quotePolicyProvider) Get(ctx context.Context, runtimeID common.Namespace, version version.Version) (*sgxQuote.Policy, error) {
+	rt, err := p.cs.Registry().GetRuntime(ctx, &registry.GetRuntimeQuery{
+		Height:           consensus.HeightLatest,
+		ID:               runtimeID,
+		IncludeSuspended: true,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to query runtime descriptor: %w", err)
+	}
+	if d := rt.DeploymentForVersion(version); d != nil {
+		var sc node.SGXConstraints
+		if err = cbor.Unmarshal(d.TEE, &sc); err != nil {
+			return nil, fmt.Errorf("malformed runtime SGX constraints: %w", err)
+		}
+		return sc.Policy, nil
+	}
+	return nil, nil
 }
